DuplicateCode->Refactoring-Step-1
Problem: Duplicate code smell is found in two classes
AddTransactionAction and AddNestedTransactionsAction, where the method
validate() is present in both classes.

Step 1: Extracted a new class TransactionValidation and pulled-up the
validate method to the new class. Made AddTransactionAction and
AddNestedTransactionsAction to extend theTransactionValidation parent
class. After this change the validate functions were tested and the
validation of the amount and accounts are working properly.