DuplicateCode->Refactoring-Step-2
Step 2: Removed the validate method from both the classes
AddNestedTransactionsAction and AddTransactionAction to eliminate
duplicate code. Then, create object for the new parent class  and the
reference of it can be used to call the validate method wherever it is
used. Here, the ValidateTxAction uses the validate method, so a
reference of TransactionValidation class can be used to implement the
alidate method. After this change the validate functions are tested and
it is working properly.