GodClass->Refactoring-Step-1
Problem: God class

Problem: The FManEntityManager acts as a god class taking to much
responsibility upon itself.It has methods to handle transactions(like
add/update/delete/get transactions),accounts,user and portfolio.

Solution: We decided to solve this code smell by extracting a separate
class that would handle all the methods that are related to Accounts.We
have created a new class FManAccountEntityManager and moved all the
methods that are responsible for account related actions to the new
class.

Step 1:Created a new Class FManAccountEntityManager and moved all the
account related methods like
addAccount,getAccount,deleteAccount,updateAccount to it.